Исправил ошибки, вернул guard в thread
